Исправлены замечания 1) в метод compareRemains передаем Values To; 2) добавлен новый объект на остатки RemainsValuesTo; 3) изменены методы compareRemains и getRemains в классе RemainsValuesPopup.
