--- src/main/java/Currency.java
@@ -1,13 +1,0 @@
-public enum Currency {
-    CURRENCYBYN("Белорусский рубль"),
-    CURRENCYUSD("Доллар США"),
-    CURRENCYEUR("Евро"),
-    CURRENCYRUB("Российский рубль");
-    private String currency;
-    Currency(String currency) {
-        this.currency = currency;
-    }
-    public String getCurrency() {
-        return currency;
-    }
-}
