+++ src/main/java/Expectations/Expectation.java
@@ -1,8 +1,12 @@
+package Expectations;
+
+import driver.SingletonDriver;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.Wait;
 import org.openqa.selenium.support.ui.WebDriverWait;
+
 import java.time.Duration;
 public class Expectation {
     WebDriver driver= SingletonDriver.getDriver();
