+++ src/main/java/assertion/MessagesProcessing.java
@@ -1,4 +1,8 @@
+package assertion;
+
+import enums.Messages;
 import org.testng.Assert;
+import pages.MainMenuPage;
 
 public class MessagesProcessing {
     public static MainMenuPage menuPage = new MainMenuPage();
