+++ src/main/java/driver/SingletonDriver.java
@@ -1,3 +1,5 @@
+package driver;
+
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
