+++ src/main/java/pages/BookOfValuesPage.java
@@ -1,7 +1,13 @@
+package pages;
+
+import Expectations.Expectation;
+import To.BookOfValuesTo;
+import factory.BookOfValuesFactory;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.testng.Assert;
+
 public class BookOfValuesPage extends PageBaseClass {
     Expectation expectation=new Expectation();
     BookOfValuesFactory bookOfValuesFactory=new BookOfValuesFactory();
