+++ src/main/java/pages/IssuanceAcceptionPage.java
@@ -1,5 +1,9 @@
+package pages;
+
+import Expectations.Expectation;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
+
 import java.math.BigDecimal;
 public class IssuanceAcceptionPage extends PageBaseClass {
     Expectation expectation=new Expectation();
