+++ src/main/java/pages/PageBaseClass.java
@@ -1,5 +1,9 @@
+package pages;
+
+import driver.SingletonDriver;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.PageFactory;
+
 public class PageBaseClass{
     WebDriver driver= SingletonDriver.getDriver();
     public PageBaseClass() {
