+++ src/main/java/pages/ReceptionOfDocsPage.java
@@ -1,6 +1,10 @@
+package pages;
+
+import Expectations.Expectation;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
+
 import java.math.BigDecimal;
 public class ReceptionOfDocsPage extends PageBaseClass {
     Expectation expectation=new Expectation();
