+++ src/main/java/properties/ConfProperties.java
@@ -1,3 +1,5 @@
+package properties;
+
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
